refactor(daily_memo): align structure with architecture
