+++ apps/daily_memo/lib/app/app.dart
@@ -1,0 +1,21 @@
+import 'package:apps.daily_memo/app/route/app_router.dart';
+import 'package:apps.daily_memo/app/route/app_routes.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class App extends StatelessWidget {
+  const App({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final router = GoRouter(
+      routes: AppRoutes.values.map((e) => e.getRouter).toList(),
+    );
+
+    return MaterialApp.router(
+      routeInformationProvider: router.routeInformationProvider,
+      routeInformationParser: router.routeInformationParser,
+      routerDelegate: router.routerDelegate,
+    );
+  }
+}
