+++ apps/daily_memo/lib/app/app_bloc_observer.dart
