+++ apps/daily_memo/lib/app/route/app_routes.dart
@@ -1,0 +1,17 @@
+enum AppRoutes {
+  home,
+  memo,
+}
+
+extension AppRoutesPath on AppRoutes {
+  String get path {
+    switch (this) {
+      case AppRoutes.home:
+        return '/';
+      case AppRoutes.memo:
+        return '/memo';
+      default:
+        return '/';
+    }
+  }
+}
