+++ apps/daily_memo/lib/features/memo/data/exceptions/external_exception.dart
@@ -1,0 +1,17 @@
+import 'package:utils/utils.dart';
+
+class ExternalException extends Failure {
+  const ExternalException({
+    super.message,
+    super.cause,
+    super.stackTrace,
+  });
+}
+
+class LocalStorageExternalException extends ExternalException {
+  const LocalStorageExternalException({
+    super.message,
+    super.cause,
+    super.stackTrace,
+  });
+}
