+++ apps/daily_memo/lib/features/memo/domain/entities/memo_info_entity.dart
