+++ apps/daily_memo/lib/features/memo/domain/usecases/add_memo_use_case.dart
@@ -1,0 +1,16 @@
+import 'package:apps.daily_memo/features/memo/domain/repositories/memo_repository.dart';
+import 'package:utils/utils.dart';
+
+class AddMemoUseCase {
+  const AddMemoUseCase({required MemoRepository repository})
+      : _repository = repository;
+
+  final MemoRepository _repository;
+
+  Future<Result<bool>> call({
+    required String title,
+    required String content,
+  }) {
+    return _repository.addMemo(title: title, content: content);
+  }
+}
