+++ apps/daily_memo/lib/features/memo/presentation/bloc/memo/memo_event.dart
