+++ apps/daily_memo/pubspec.yaml
@@ -26,6 +26,8 @@ dependencies:
   equatable: ^2.0.3
   flutter_bloc: ^8.1.1
   get_it: ^7.6.4
+  utils:
+    path: ../../packages/utils
 
 dev_dependencies:
   flutter_test:
@@ -37,4 +39,3 @@ dev_dependencies:
 
 flutter:
   uses-material-design: true
-
